Added freetype, added new project for editor and moving away from Sandbox. Also added framebuffers and uniform buffers for scene camera data. Will start work on text renderering API soon.

+++ Ruby/src/API/OpenGL/OpenGLVertexArray.cpp
@@ -67,7 +67,7 @@ namespace Ruby {
 			default:
 				RB_ASSERT(false, "Unsupported data type!");
 			}
-			offset += e[i].count * sizeOfLayoutType(e[i].type);
+			offset += (uint64_t)e[i].count * (uint64_t)sizeOfLayoutType(e[i].type);
 		}
 
 		m_VertexBuffers.push_back(buffer);
